created partition wrapper
